+++ lib/config/configuration.js
@@ -1,13 +1,14 @@
 'use strict';
 
-const { resolvePathFor } = require('./utils');
+const { resolvePathFor } = require('../utils');
+const defaultConfiguration = require('./default_configuration');
 const FailFast = require('./fail_fast');
 const CONFIGURATION_FILE_NAME = '.testyrc.json';
 
 class Configuration {
-  
+
   // instance creation
-  
+
   static current() {
     let userConfiguration;
     try {
@@ -15,22 +16,22 @@ class Configuration {
     } catch (error) {
       userConfiguration = {};
     }
-    const defaultConfiguration = require('./default_configuration');
+
     return new this(userConfiguration, defaultConfiguration);
   }
-  
+
   constructor(userConfiguration, defaultConfiguration) {
     this._userConfigurationOptions = userConfiguration;
     this._defaultConfigurationOptions = defaultConfiguration;
     this._initializeOptions();
   }
-  
+
   // accessing configuration parameters
-  
+
   directory() {
     return this._configurationOptions.directory;
   }
-  
+
   filter() {
     return new RegExp(this.filterRaw());
   }
@@ -38,32 +39,32 @@ class Configuration {
   filterRaw() {
     return this._configurationOptions.filter;
   }
-  
+
   language() {
     return this._configurationOptions.language;
   }
-  
+
   failFastMode() {
     return new FailFast(this._configurationOptions.failFast);
   }
-  
+
   randomOrder() {
     return this._configurationOptions.randomOrder;
   }
-  
+
   // configuring
-  
+
   configureUI(ui) {
     ui.useLanguage(this.language());
   }
-  
+
   configureTestRunner(testRunner) {
     testRunner.setFailFastMode(this.failFastMode());
     testRunner.setTestRandomness(this.randomOrder());
   }
-  
+
   // initialization
-  
+
   _initializeOptions() {
     this._configurationOptions = { ...this._defaultConfigurationOptions, ...this._userConfigurationOptions };
   }
