+++ lib/config/default_configuration.json
