+++ lib/config/fail_fast.js
