+++ lib/core/asserter.js
@@ -3,9 +3,9 @@
 const TestResult = require('./test_result');
 const TestResultReporter = require('./test_result_reporter');
 const Assertion = require('./assertion');
-const { I18nMessage } = require('./i18n');
+const { I18nMessage } = require('../i18n/i18n');
 
-const { isStringWithContent } = require('./utils');
+const { isStringWithContent } = require('../utils');
 
 class FailureGenerator extends TestResultReporter {
   /**
