+++ lib/core/assertion.js
@@ -3,9 +3,9 @@
 const TestResultReporter = require('./test_result_reporter');
 const EqualityAssertionStrategy = require('./equality_assertion_strategy');
 const IdentityAssertionStrategy = require('./identity_assertion_strategy');
-const { I18nMessage } = require('./i18n');
+const { I18nMessage } = require('../i18n/i18n');
 
-const { prettyPrint, isUndefined, isRegex, notNullOrUndefined, numberOfElements, convertToArray } = require('./utils');
+const { prettyPrint, isUndefined, isRegex, notNullOrUndefined, numberOfElements, convertToArray } = require('../utils');
 
 /**
  * I represent an assertion we want to make on a specific object (called the `actual`), against an expectation, in the
