+++ lib/core/equality_assertion_strategy.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const { I18nMessage } = require('./i18n');
-const { isString, isFunction, isUndefined, respondsTo, isCyclic, deepStrictEqual } = require('./utils');
+const { I18nMessage } = require('../i18n/i18n');
+const { isString, isFunction, isUndefined, respondsTo, isCyclic, deepStrictEqual } = require('../utils');
 
 const EqualityAssertionStrategy = {
   availableStrategies() {
