+++ lib/core/identity_assertion_strategy.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const { I18nMessage } = require('./i18n');
-const { isUndefined } = require('./utils');
+const { I18nMessage } = require('../i18n/i18n');
+const { isUndefined } = require('../utils');
 
 const IdentityAssertionStrategy = {
   availableStrategies() {
@@ -10,7 +10,7 @@ const IdentityAssertionStrategy = {
       DefaultIdentity,
     ];
   },
-  
+
   evaluate(actual, expected) {
     return this.availableStrategies()
       .find(strategy => strategy.appliesTo(actual, expected))
@@ -20,11 +20,11 @@ const IdentityAssertionStrategy = {
 
 const BothPartsUndefined = {
   __proto__: IdentityAssertionStrategy,
-  
+
   appliesTo(actual, expected) {
     return isUndefined(actual) && isUndefined(expected);
   },
-  
+
   evaluate() {
     return {
       comparisonResult: undefined,
@@ -36,11 +36,11 @@ const BothPartsUndefined = {
 
 const DefaultIdentity = {
   __proto__: IdentityAssertionStrategy,
-  
+
   appliesTo(_actual, _expected) {
     return true;
   },
-  
+
   evaluate(actual, expected) {
     return {
       comparisonResult: actual === expected,
