+++ lib/core/test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const TestResult = require('./test_result');
-const { isStringWithContent, isFunction, isUndefined } = require('./utils');
+const { isStringWithContent, isFunction, isUndefined } = require('../utils');
 
 /**
  * I am an executable test, part of a {@link TestSuite} and executed by a {@link TestRunner}.
