+++ lib/core/test_result.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { detectUserCallingLocation } = require('./utils');
+const { detectUserCallingLocation } = require('../utils');
 
 class TestResult {
   // instance creation messages
