+++ lib/core/test_result_reporter.js
