+++ lib/core/test_runner.js
@@ -2,8 +2,8 @@
 
 const Test = require('./test');
 const TestSuite = require('./test_suite');
-const FailFast = require('./fail_fast');
-const { shuffle } = require('./utils');
+const FailFast = require('../config/fail_fast');
+const { shuffle } = require('../utils');
 
 /**
  * I am responsible for executing [test suites]{@link TestSuite}, collecting the results of each test of each suite
