+++ lib/core/test_suite.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { shuffle, isUndefined, isStringWithContent, isFunction } = require('./utils');
+const { shuffle, isUndefined, isStringWithContent, isFunction } = require('../utils');
 
 class TestSuite {
   static BEFORE_HOOK_NAME = 'before';
