+++ lib/i18n/i18n.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const TRANSLATIONS = require('./translations');
-const { subclassResponsibility, isUndefined } = require('./utils');
+const { subclassResponsibility, isUndefined } = require('../utils');
 
 class I18nMessage {
   static of(key, ...params) {
