+++ lib/script_action.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { Testy } = require('./testy');
-const Configuration = require('./configuration');
+const Configuration = require('./config/configuration');
 
 const ScriptAction = {
   availableActions() {
