+++ lib/testy.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const TestRunner = require('./test_runner');
-const { Asserter, FailureGenerator, PendingMarker } = require('./asserter');
-const ConsoleUI = require('./console_ui');
+const TestRunner = require('./core/test_runner');
+const { Asserter, FailureGenerator, PendingMarker } = require('./core/asserter');
+const ConsoleUI = require('./ui/console_ui');
 const { allFilesMatching, resolvePathFor, errorDetailOf } = require('./utils');
-const { I18nMessage } = require('./i18n');
+const { I18nMessage } = require('./i18n/i18n');
 
 const ui = new ConsoleUI(process, console);
 const testRunner = new TestRunner(ui.testRunnerCallbacks());
@@ -62,6 +62,10 @@ const pending = new PendingMarker(testRunner);
 function test(name, testBody) {
   testRunner.registerTest(name, testBody, ui.testCallbacks());
 }
+
+test.only = (name, testBody) => {
+  testRunner.registerTest(name, testBody, ui.testCallbacks());
+};
 
 /**
  * Defines a new test suite. Suites are expected to define [tests]{@link test} inside it.
