+++ lib/ui/console_ui.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { I18n } = require('./i18n');
+const { I18n } = require('../i18n/i18n');
 const Formatter = require('./formatter');
 
 class ConsoleUI {
