+++ lib/ui/formatter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { isString, isEmpty } = require('./utils');
+const { isString, isEmpty } = require('../utils');
 
 // Colors and emphasis
 const off = '\x1b[0m';
