+++ tests/core/assertions/exception_assertions_test.js
@@ -4,7 +4,7 @@ const { suite, test } = require('../../../lib/testy');
 const { resultOfATestWith } = require('../../support/runner_helpers');
 const { expectSuccess, expectFailureOn } = require('../../support/assertion_helpers');
 
-const { I18nMessage } = require('../../../lib/i18n');
+const { I18nMessage } = require('../../../lib/i18n/i18n');
 
 suite('exception assertions', () => {
   test('raises() can receive a string and it passes when the exact string is expected', async() => {
