+++ tests/core/assertions/numeric_assertions_test.js
@@ -4,7 +4,7 @@ const { suite, test } = require('../../../lib/testy');
 const { resultOfATestWith } = require('../../support/runner_helpers');
 const { expectSuccess, expectFailureOn } = require('../../support/assertion_helpers');
 
-const { I18nMessage } = require('../../../lib/i18n');
+const { I18nMessage } = require('../../../lib/i18n/i18n');
 
 suite('numeric assertions', () => {
   test('isNearTo passes if an exact integer matches', async() => {
