+++ tests/core/i18n_messages_test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { suite, test, assert } = require('../../lib/testy');
-const { I18n, I18nMessage } = require('../../lib/i18n');
+const { I18n, I18nMessage } = require('../../lib/i18n/i18n');
 
 suite('i18n messages', () => {
   const translations = { en: { key1: 'value 1', key2: 'value 2', key3: 'value 3' } };
