+++ tests/core/test_test.js
@@ -5,8 +5,8 @@ const { aTestWithNoAssertions, aTestWithBody } = require('../support/tests_facto
 const { withRunner, runSingleTest, resultOfASuiteWith } = require('../support/runner_helpers');
 const { expectErrorOn, expectFailureOn } = require('../support/assertion_helpers');
 
-const Test = require('../../lib/test');
-const { I18nMessage } = require('../../lib/i18n');
+const Test = require('../../lib/core/test');
+const { I18nMessage } = require('../../lib/i18n/i18n');
 
 suite('tests behavior', () => {
   const noop = () => {
