+++ tests/support/tests_factory.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Test = require('../../lib/test');
+const Test = require('../../lib/core/test');
 
 const noop = () => {
   // intentional empty function
